requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.35
openpyxl==3.1.2
requests==2.31.0
PyJWT==2.8.0
python-dateutil==2.8.2